eliminados archivos playbutton, backbutton, ranckbutton, instrbutton, exitbutton, y movidas la clases a un solo archivo buttons
