--- BackButton.cpp
@@ -1,19 +1,0 @@
-#include "BackButton.h"
-
-BackButton::BackButton(int x_position, int y_position, Image* up, Image* down,RosalilaGraphics* painter): Button(x_position,y_position,up,down,painter)
-{
-    //ctor
-}
-
-void BackButton:: funcionamiento(){
-
-
-}
-BackButton::BackButton(){
-
-}
-
-BackButton::~BackButton()
-{
-    //dtor
-}
