+++ Demo.cbp
@@ -34,38 +34,28 @@
 			<Add option="-Wall" />
 			<Add option="-fexceptions" />
 		</Compiler>
-		<Unit filename="BackButton.cpp" />
-		<Unit filename="BackButton.h" />
 		<Unit filename="Button.cpp" />
 		<Unit filename="Button.h" />
 		<Unit filename="Buttons.cpp" />
 		<Unit filename="Buttons.h" />
 		<Unit filename="Car.cpp" />
 		<Unit filename="Car.h" />
-		<Unit filename="ExitButton.cpp" />
-		<Unit filename="ExitButton.h" />
 		<Unit filename="Font.cpp" />
 		<Unit filename="Font.h" />
 		<Unit filename="Game.cpp" />
 		<Unit filename="Game.h" />
 		<Unit filename="InsScreen.cpp" />
 		<Unit filename="InsScreen.h" />
-		<Unit filename="InstrButton.cpp" />
-		<Unit filename="InstrButton.h" />
 		<Unit filename="ListImage.cpp" />
 		<Unit filename="ListImage.h" />
 		<Unit filename="MenuScreen.cpp" />
 		<Unit filename="MenuScreen.h" />
 		<Unit filename="Pista.cpp" />
 		<Unit filename="Pista.h" />
-		<Unit filename="PlayButton.cpp" />
-		<Unit filename="PlayButton.h" />
 		<Unit filename="PlayScreen.cpp" />
 		<Unit filename="PlayScreen.h" />
 		<Unit filename="RaceGear.cpp" />
 		<Unit filename="RaceGear.h" />
-		<Unit filename="RankButton.cpp" />
-		<Unit filename="RankButton.h" />
 		<Unit filename="RankingScreen.cpp" />
 		<Unit filename="RankingScreen.h" />
 		<Unit filename="Rosalila/RosalilaAI/Automata.cpp" />
